+++ types/components.ts
@@ -1,4 +1,9 @@
-import { IButton } from './misc'
+export interface IButton {
+  text: string
+  to: string
+  /** If true, will use href instead of to attribute for navigating */
+  external?: boolean
+}
 
 export interface ICarouselSlide {
   id: number
@@ -16,4 +21,20 @@ export interface IIconCard {
   title: string
   description: string
   btn?: IButton
+}
+
+export interface IFeature {
+  id: number
+  src?: string
+  title: string
+  description: string
+  btn?: IButton
+}
+
+export interface IImageBlock {
+  id: number
+  src: string
+  title: string
+  titleMuted?: string
+  description: string
 }
